requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2
openai==1.3.7
requests==2.32.3
playwright==1.48.0
tenacity==9.0.0
python-dotenv==1.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
httpx==0.25.1
pdf2image==1.17.0
